refactor(combo): extract shared types to combo/types.ts
